Smooth Scrolling for Filemanager Lighttable
The file manager in the lighttable scrolls only in one-thumbnail increments. Thus feels clunky and disorienting compared to the smooth scrolling that is available if one drags the scroll bars.

This commit changes the file manager scroll behavior to scroll pixel-precise if scrolled with a precision touch pad. Normal "clicky" mouse wheel scroll behavior is unchanged.

The commit changes scroll events to be either, 1.0 or -1.0 from clicky mouse wheels, or any other floating point offset for precision touch pads. If a 1.0 or -1.0 is detected, the view is scrolled by full-thumbnail increments. If any other offset is detected, it is scrolled by that pixel amount.