Smooth scrolling preference replaces heuristic
previously, the lighttable either scrolled smoothly if a precision scroll event was registered (scroll value is != 1 or -1), or row-by-row if it was clicky (scroll value is 1 or -1). This former is typically issued by touch pads, where smooth scrolling is common behavior. The latter for clicky mouse scroll wheels with clear detents.

However, it appears that some mice with clicky wheels still issue floating scroll events, and smooth scrolling. This behavior did not feel good.

Therefore, this commit makes the choice between smooth scrolling and clicky scrolling a preference instead, and no longer tries to infer "correct" scrolling behavior from a scroll increment heuristic.